requirements.txt
flask>=3.0
pypdf>=4.0
pandas>=2.0
openpyxl>=3.1
gunicorn>=21.0
psycopg2-binary>=2.9
